+++ fstd2nc/mixins/xycoords.py
@@ -402,8 +402,8 @@ class PolarStereo(GridMap):
       pole = gdxyfll (self._grd['id'], 90, 0)
     else:
       pole = gdxyfll (self._grd['id'], -90, 0)
-    px = np.rint(pole['x'][0] - 1) * self._res
-    py = np.rint(pole['y'][0] - 1) * self._res
+    px = np.rint(pole['x'][0] - np.float64(1.0)) * self._res
+    py = np.rint(pole['y'][0] - np.float64(1.0)) * self._res
     self._false_easting =  px - self._ax[0]
     self._false_northing = py - self._ay[0]
     self._xaxisatts['long_name'] = 'x-coordinate of polar-stereographic projection'
